showing modal on label click
